requirements.txt
python-dotenv==1.0.1
aiogram==3.4.1
httpx==0.27.0
pydantic==2.6.4